added xml-format
This version improves the previous in several ways:
Add new user option xml-format-declaration to handle the remove/insert of XML declaration.
Add new function xml-declaration-p.
Add new function and key xml-remove-declaration.
xml-format now keep the position and mark, handle the XML declaration and add a new line when region
does not begin on the beginning of the line.